Restructured app to run as a Python package using Blueprints and Factories. Closes #18
