--- webapp/gunicorn.py
@@ -1,7 +1,0 @@
-# gunicorn.py
-import os
-
-if os.environ.get('MODE') == 'dev':
-    reload = True
-
-bind = '0.0.0.0:5000'
