--- webapp/instance/flask.cfg
@@ -1,20 +1,0 @@
-import os
-
-
-# grab the folder of the top-level directory of this project
-BASEDIR = os.path.abspath(os.path.dirname(__file__))
-TOP_LEVEL_DIR = os.path.abspath(os.curdir)
-
-# Update later by using a random number generator and moving
-# the actual key outside of the source code under version control
-SECRET_KEY = 'secret_key'
-WTF_CSRF_ENABLED = True
-DEBUG = True
-
-# SQLAlchemy
-MYSQL_DEFAULT_USER = 'postgres'
-MYSQL_USER = 'mpower'
-MYSQL_PASSWORD = 'mpower'
-MYSQL_DATABASE = 'mpower_api'
-SQLALCHEMY_DATABASE_URI = 'mysql+pymysql://' + MYSQL_USER + ':' + MYSQL_PASSWORD + '@db:3306/' + MYSQL_DB
-SQLALCHEMY_TRACK_MODIFICATIONS = True
