+++ webapp/mpowerapi/mpowerapi/blueprints/static.py
@@ -1,5 +1,3 @@
-from app import app
-
 @app.route('/')
 def index():
     return "Hello, World!"
